+++ core/src/stored/stored_conf.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2009 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -290,11 +290,9 @@ static void StoreAutopassword(LEX* lc, ResourceItem* item, int index, int pass)
 {
   switch ((*item->allocated_resource)->rcode_) {
     case R_DIRECTOR:
-      /*
-       * As we need to store both clear and MD5 hashed within the same
+      /* As we need to store both clear and MD5 hashed within the same
        * resource class we use the item->code as a hint default is 0
-       * and for clear we need a code of 1.
-       */
+       * and for clear we need a code of 1. */
       switch (item->code) {
         case 1:
           my_config->StoreResource(CFG_TYPE_CLEARPASSWORD, lc, item, index,
@@ -491,10 +489,8 @@ static void CheckDropletDevices(ConfigurationParser& my_config)
     DeviceResource* d = dynamic_cast<DeviceResource*>(p);
     if (d && d->device_type == DeviceType::B_DROPLET_DEV) {
       if (d->max_concurrent_jobs == 0) {
-        /*
-         * 0 is the general default. However, for this device_type, only 1
-         * works. So we set it to this value.
-         */
+        /* 0 is the general default. However, for this device_type, only 1
+         * works. So we set it to this value. */
         Jmsg1(nullptr, M_WARNING, 0,
               _("device %s is set to the default 'Maximum Concurrent Jobs' = "
                 "1.\n"),
@@ -737,9 +733,6 @@ static bool SaveResource(int type, ResourceItem* items, int pass)
   int i;
   int error = 0;
 
-  BareosResource* allocated_resource = *resources[type].allocated_resource_;
-  if (pass == 2 && !allocated_resource->Validate()) { return false; }
-
   // Ensure that all required items are present
   for (i = 0; items[i].name; i++) {
     if (items[i].flags & CFG_ITEM_REQUIRED) {
@@ -758,6 +751,9 @@ static bool SaveResource(int type, ResourceItem* items, int pass)
 
   // save previously discovered pointers into dynamic memory
   if (pass == 2) {
+    BareosResource* allocated_resource = my_config->GetResWithName(
+        type, (*items->allocated_resource)->resource_name_);
+    if (allocated_resource && !allocated_resource->Validate()) { return false; }
     switch (type) {
       case R_DEVICE:
       case R_MSGS:
@@ -765,8 +761,8 @@ static bool SaveResource(int type, ResourceItem* items, int pass)
         // Resources not containing a resource
         break;
       case R_DIRECTOR: {
-        DirectorResource* p = dynamic_cast<DirectorResource*>(
-            my_config->GetResWithName(R_DIRECTOR, res_dir->resource_name_));
+        DirectorResource* p
+            = dynamic_cast<DirectorResource*>(allocated_resource);
         if (!p) {
           Emsg1(M_ERROR_TERM, 0, _("Cannot find Director resource %s\n"),
                 res_dir->resource_name_);
@@ -777,8 +773,7 @@ static bool SaveResource(int type, ResourceItem* items, int pass)
         break;
       }
       case R_STORAGE: {
-        StorageResource* p = dynamic_cast<StorageResource*>(
-            my_config->GetResWithName(R_STORAGE, res_store->resource_name_));
+        StorageResource* p = dynamic_cast<StorageResource*>(allocated_resource);
         if (!p) {
           Emsg1(M_ERROR_TERM, 0, _("Cannot find Storage resource %s\n"),
                 res_store->resource_name_);
@@ -793,9 +788,7 @@ static bool SaveResource(int type, ResourceItem* items, int pass)
       }
       case R_AUTOCHANGER: {
         AutochangerResource* p
-            = dynamic_cast<AutochangerResource*>(my_config->GetResWithName(
-                R_AUTOCHANGER, res_changer->resource_name_));
-
+            = dynamic_cast<AutochangerResource*>(allocated_resource);
         if (!p) {
           Emsg1(M_ERROR_TERM, 0, _("Cannot find AutoChanger resource %s\n"),
                 res_changer->resource_name_);